Create project for Neo.ConsoleService
