+++ Neo.ConsoleService/CommandSpaceToken.cs
@@ -1,10 +1,10 @@
 using System;
 using System.Diagnostics;
 
-namespace Neo.CommandParser
+namespace Neo.ConsoleService
 {
     [DebuggerDisplay("Value={Value}, Count={Count}")]
-    public class CommandSpaceToken : CommandToken
+    internal class CommandSpaceToken : CommandToken
     {
         /// <summary>
         /// Count
