+++ Neo.ConsoleService/CommandTokenType.cs
@@ -1,0 +1,9 @@
+namespace Neo.ConsoleService
+{
+    internal enum CommandTokenType : byte
+    {
+        String,
+        Space,
+        Quote,
+    }
+}
