--- neo-cli/CommandParser/CommandTokenType.cs
@@ -1,9 +1,0 @@
-namespace Neo.CommandParser
-{
-    public enum CommandTokenType : byte
-    {
-        String,
-        Space,
-        Quote,
-    }
-}
